verisql/requirements.txt
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
z3-solver>=4.12.0
sqlglot>=20.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
